+++ packages/airless-core/CHANGELOG.md
@@ -1,5 +1,6 @@
 
 **unreleased**
+- [Feature] Add methods to `DatalakeHook` to normalize data before storing to datalake
 
 **v0.2.0**
 - [Deprecation] Remove deprecated class `BaseDto`